+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { VerificationProviderController } from './modules/verification-provider/controllers/verification-provider.controller';
-import { VerificationProviderService } from './modules/verification-provider/services/verification-provider.service';
+import { VerificationProviderController } from './modules/verification-provider/verification-provider.controller';
+import { VerificationProviderService } from './modules/verification-provider/verification-provider.service';
 
 
 @Module({
